addressing issue #5 continued, move cobra_remote_service class into its own file
